embeddings-clusters/requirements.txt
python-dotenv
requests
pandas
umap-learn
numpy
scikit-learn
seaborn
matplotlib
hdbscan
plotly
nbformat
POT
